// packages/disk-moon-city/moonCity/types.ts
export enum STORAGE_TYPE {
  NONE = 'none',
  CREDITS = 'credits',
  MINERAL_ORE = 'mineral_ore',
  ENERGY_CELL = 'energy_cell',
  FOOD = 'food',
  ENERGY = 'energy',
  HUMAN = 'human',
  EMPLOYEE = 'employee',
  SECURITY_SERVICE = 'security_service',
  SOLDIER = 'soldiers',
  MERCENARY = 'mercenary',
  ENERGY_TRANSFER = 'energy_transfer',
  SHIELD_ENERGY = 'shield_energy'
}
// ...
export enum LINE_GROUP {
  VEHICLE = 'vehicle',
  GENERAL = 'general',
  COST = 'cost',
  INCOME = 'income',
  ATTACK = 'attack'
}